refactor: storybook 수정
