+++ src/pages/CardListPage/index.stories.jsx
@@ -1,14 +1,10 @@
 import { BrowserRouter } from 'react-router-dom'
 import CardListPage from 'pages/CardListPage'
 
-import { CardInfoProvider } from 'context/cardInfo-context'
-
 const Template = () => {
   return (
     <BrowserRouter>
-      <CardInfoProvider>
-        <CardListPage />
-      </CardInfoProvider>
+      <CardListPage />
     </BrowserRouter>
   )
 }
